stats-service added
