+++ sass/main/--result/main--result.scss
@@ -30,3 +30,13 @@
 .main-replay:hover {
   color: red;
 }
+
+.server-upload {
+  position: absolute;
+  top: 50px;
+  left: 50px;
+  color: black;
+  background-color: #ff9749;
+  padding: 5px 10px;
+  border-radius: 5px;
+}
